Piped
Pipe script is able to take in num_ind and generate the num_ind of initial population.
It then generates a round robin list of fightes and puts all individuals up against each other.
It then is able to generate a df and update the ELO Ratings respectively after each match